+++ package.json
@@ -18,7 +18,7 @@
     "react-dom": "^16.12.0",
     "react-scripts": "^3.3.1",
     "react-use": "^13.24.0",
-    "typescript": "^3.7.3"
+    "typescript": "^3.8.1-rc"
   },
   "scripts": {
     "start": "react-scripts start",
